demo/requirements.txt
# python 3.12
Django==5.2.7
django-environ
pandas
# django_gauth==0.2.0
django_gauth==0.2.0
google-api-python-client
# gsheet_tools==0.2.0
gsheet_tools==0.2.0
icecream
